chore: Enable contract to be used as a dependency and add error module (#10)
This PR introduces some improvements recommended by the [CosmWasm
book](https://book.cosmwasm.com/). These are multiple bullets but they
are all very small things
1. Move all entry point functions to the `src/lib.rs` module as wrappers
to the functions implemented by the `src/contract.rs` module. This
enables our contract to be used as a dependency by other contracts
without it being messed up by the `#[entry_point]` macros. This is
described in detail
[here](https://book.cosmwasm.com/basics/good-practices.html#disabling-entry-points-for-libraries)
2. Create the `src/error.rs` module to wrap our contract related errors.
This is a nice addition which enables us to manage errors with the
`thiserror` package. This recommendation is described
[here](https://book.cosmwasm.com/basics/execute.html#error-handling)
3. Namespace query handling functions under a `mod query { ... }` block.
This allows us to avoid having the prefix `query` for everything.
Recommended
[here](https://book.cosmwasm.com/basics/query.html#improving-the-message)
4. Add aliases for common functionality, i.e. `cargo schema`, `cargo
wasm`, and `cargo wasm-debug`.
5. Add a `make` command for building for an optimized version for
`arm64` architectures.
6. Removed the contents of the reply method as we don't invoke any
messages or submessages from the contract so there's nothing to reply
to.
7. Moved ibc related methods to a dedicated `ibc.rs`.
8. Introduced `cw-multi-test` package and used it for integration tests.
This allows us to test a contract end-to-end instead of calling our
internal methods, enabling easier testing. See the explanation
[here](https://book.cosmwasm.com/basics/multitest-intro.html#introducing-multitest)
9. Introduce an empty `execute` message. Mostly needed for the tests,
but I figured we should have a placeholder as we'll have messages later.

Also, I added some comments and TODOs posing some questions:
a. The code doesn't use the `reflect` contract at the moment, but we
have some leftovers relating to the reflect contract ID etc. This is
some redundant state. Do we want to use the `reflect` contract? If so,
let's fully support it, and if not, let's remove references.
b. The `ibc_packet_ack` and `ibc_packet_timeout` methods should not be
called (as this contract only receives packets). Should we error or
panic there?
c. We do not store any information about the channels that are opened
etc. How should we handle those? Who can open channels?

Very new to Rust and smart contracts in general so thoughts appreciated!