crypto: use k256 to implement adaptor sig (#43)
Resolves #13 

This PR replaces the usage of `schnorr_fun` and `secp256kfun` with
`k256` (which is native in CosmWasm) for implementing the adaptor
signature verification. This gets rid of `schnorr_fun` and `secp256kfun`
which can bloat the contract size.

This PR also runs `cargo fix-lint` to lint the entire project.