feat: Remove slashing evidence message
There's no benefit for someone directly submitting slashing evidence to the CosmWasm
smart contract, especially given that it is submitted on Babylon Genesis which
has a double signing evidence handler of its own.

The commit removes the sub message for handling slashing evidence and in the case
of a detected double-signing, forwards an evidence message to the underlying
Babylon Genesis chain directly.