feat: rate limiting for FP messages (#80)
## Description

<!-- Brief description of changes -->

This PR introduces rate limiting on the number of msgs that a FP can
submit within each time interval. This includes

- config for rate limiting, including the interval and the max number of
msgs per interval
- extending instantiation msg for supplying rate limiting config
- rate limiting KV store. It is a simple counter where key is FP PK and
value is a tuple of interval number and count. The interval number is
determined by height / interval
- a single function for checking rate limiting and accumulating the
counter
- checking rate limiting for both pub rand commit and finality sig

This PR also provides unit tests for instantiation msg and rate
limiting, and updates e2e tests.

Resolves #77

## Checklist

- [x] I have updated the
[SPEC.md](https://github.com/babylonlabs-io/rollup-bsn-contracts/blob/main/SPEC.md)
file if this change affects the specification
- [x] I have updated the schema by running `cargo gen-schema`